Split write unit tests
