+++ test/unit/operations/write/mergePayload.ts
@@ -1,0 +1,78 @@
+import * as util from 'util';
+
+import { CacheContext } from '../../../../src/context';
+import { GraphSnapshot } from '../../../../src/GraphSnapshot';
+import { write } from '../../../../src/operations/write';
+import { NodeId, StaticNodeId } from '../../../../src/schema';
+import { query } from '../../../helpers';
+
+const { QueryRoot: QueryRootId } = StaticNodeId;
+
+// These are really more like integration tests, given the underlying machinery.
+//
+// It just isn't very fruitful to unit test the individual steps of the write
+// workflow in isolation, given the contextual state that must be passed around.
+describe(`operations.write`, () => {
+  const config = new CacheContext({
+    logger: {
+      warn(message: string, ...args: any[]) {
+        throw new Error(util.format(message, ...args));
+      },
+    },
+  });
+
+  const rootValuesQuery = query(`{ foo bar }`);
+
+  const empty = new GraphSnapshot();
+
+  describe(`merge unidentifiable payloads with previously known nodes`, () => {
+    let baseline: GraphSnapshot, snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      const baselineResult = write(config, empty, rootValuesQuery, {
+        foo: { id: 1, name: 'Foo' },
+        bar: { id: 2, name: 'Bar' },
+      });
+      baseline = baselineResult.snapshot;
+
+      const result = write(config, baseline, rootValuesQuery, {
+        foo: { name: 'Foo Boo' },
+        bar: { extra: true },
+      });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`doesn't mutate the previous versions`, () => {
+      expect(baseline.get(QueryRootId)).to.not.eq(snapshot.get(QueryRootId));
+      expect(baseline.get('1')).to.not.eq(snapshot.get('1'));
+      expect(baseline.get('2')).to.not.eq(snapshot.get('2'));
+      expect(baseline.get(QueryRootId)).to.deep.eq({
+        foo: { id: 1, name: 'Foo' },
+        bar: { id: 2, name: 'Bar' },
+      });
+    });
+
+    it(`updates existing values in referenced nodes`, () => {
+      expect(snapshot.get('1')).to.deep.eq({ id: 1, name: 'Foo Boo' });
+    });
+
+    it(`inserts new values in referenced nodes`, () => {
+      expect(snapshot.get('2')).to.deep.eq({ id: 2, name: 'Bar', extra: true });
+    });
+
+    it(`updates references to the newly edited nodes`, () => {
+      const root = snapshot.get(QueryRootId);
+      expect(root.foo).to.eq(snapshot.get('1'));
+      expect(root.bar).to.eq(snapshot.get('2'));
+    });
+
+    it(`doesn't mark regenerated nodes as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members(['1', '2']);
+    });
+
+    it(`contains the correct nodes`, () => {
+      expect(snapshot.allNodeIds()).to.have.members([QueryRootId, '1', '2']);
+    });
+
+  });
+});
