+++ test/unit/operations/write/orphaneNode.ts
@@ -1,0 +1,202 @@
+import * as util from 'util';
+
+import { CacheContext } from '../../../../src/context';
+import { GraphSnapshot } from '../../../../src/GraphSnapshot';
+import { write } from '../../../../src/operations/write';
+import { NodeId, StaticNodeId } from '../../../../src/schema';
+import { query } from '../../../helpers';
+
+const { QueryRoot: QueryRootId } = StaticNodeId;
+
+// These are really more like integration tests, given the underlying machinery.
+//
+// It just isn't very fruitful to unit test the individual steps of the write
+// workflow in isolation, given the contextual state that must be passed around.
+describe(`operations.write`, () => {
+  const config = new CacheContext({
+    logger: {
+      warn(message: string, ...args: any[]) {
+        throw new Error(util.format(message, ...args));
+      },
+    },
+  });
+
+  const rootValuesQuery = query(`{ foo bar }`);
+
+  const empty = new GraphSnapshot();
+
+  describe(`when orphaning a node`, () => {
+    let baseline: GraphSnapshot, snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      const baselineResult = write(config, empty, rootValuesQuery, {
+        foo: { id: 1, name: 'Foo' },
+        bar: { id: 2, name: 'Bar' },
+      });
+      baseline = baselineResult.snapshot;
+
+      const result = write(config, baseline, rootValuesQuery, {
+        bar: null,
+      });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`doesn't mutate the previous versions`, () => {
+      expect(baseline.get(QueryRootId)).to.not.eq(snapshot.get(QueryRootId));
+      expect(baseline.get('1')).to.eq(snapshot.get('1'));
+      expect(baseline.get('2')).to.not.eq(snapshot.get('2'));
+      expect(baseline.get(QueryRootId)).to.deep.eq({
+        foo: { id: 1, name: 'Foo' },
+        bar: { id: 2, name: 'Bar' },
+      });
+    });
+
+    it(`replaces the reference with null`, () => {
+      expect(snapshot.get(QueryRootId)).to.deep.eq({
+        foo: { id: 1, name: 'Foo' },
+        bar: null,
+      });
+    });
+
+    it(`updates outbound references`, () => {
+      const queryRoot = snapshot.getNodeSnapshot(QueryRootId)!;
+      expect(queryRoot.outbound).to.have.deep.members([{ id: '1', path: ['foo'] }]);
+    });
+
+    it(`marks the container and orphaned node as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members([QueryRootId, '2']);
+    });
+
+    it(`contains the correct nodes`, () => {
+      expect(snapshot.allNodeIds()).to.have.members([QueryRootId, '1']);
+    });
+
+  });
+
+  describe(`orphane a node`, () => {
+
+    let baseline: GraphSnapshot, snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      const baselineResult = write(config, empty, rootValuesQuery, {
+        foo: { id: 1, name: 'Foo' },
+        bar: { id: 2, name: 'Bar' },
+      });
+      baseline = baselineResult.snapshot;
+
+      const result = write(config, baseline, rootValuesQuery, {
+        bar: null,
+      });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`doesn't mutate the previous versions`, () => {
+      expect(baseline.get(QueryRootId)).to.not.eq(snapshot.get(QueryRootId));
+      expect(baseline.get('1')).to.eq(snapshot.get('1'));
+      expect(baseline.get('2')).to.not.eq(snapshot.get('2'));
+      expect(baseline.get(QueryRootId)).to.deep.eq({
+        foo: { id: 1, name: 'Foo' },
+        bar: { id: 2, name: 'Bar' },
+      });
+    });
+
+    it(`replaces the reference with null`, () => {
+      expect(snapshot.get(QueryRootId)).to.deep.eq({
+        foo: { id: 1, name: 'Foo' },
+        bar: null,
+      });
+    });
+
+    it(`updates outbound references`, () => {
+      const queryRoot = snapshot.getNodeSnapshot(QueryRootId)!;
+      expect(queryRoot.outbound).to.have.deep.members([{ id: '1', path: ['foo'] }]);
+    });
+
+    it(`marks the container and orphaned node as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members([QueryRootId, '2']);
+    });
+
+    it(`contains the correct nodes`, () => {
+      expect(snapshot.allNodeIds()).to.have.members([QueryRootId, '1']);
+    });
+  });
+
+  describe(`orphane a subgraph`, () => {
+
+    let baseline: GraphSnapshot, snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      const baselineResult = write(config, empty, rootValuesQuery, {
+        foo: {
+          id: 1,
+          name: 'Foo',
+          two: { id: 222 },
+        },
+        bar: {
+          id: 2,
+          one: { id: 111 },
+          two: { id: 222 },
+          three: {
+            id: 333,
+            foo: { id: 1 },
+          },
+        },
+      });
+      baseline = baselineResult.snapshot;
+
+      const result = write(config, baseline, rootValuesQuery, {
+        foo: { two: null },
+        bar: null,
+      });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`doesn't mutate the previous versions`, () => {
+      expect(baseline.get(QueryRootId)).to.deep.eq({
+        foo: {
+          id: 1,
+          name: 'Foo',
+          two: { id: 222 },
+        },
+        bar: {
+          id: 2,
+          one: { id: 111 },
+          two: { id: 222 },
+          three: {
+            id: 333,
+            foo: {
+              id: 1,
+              name: 'Foo',
+              two: { id: 222 },
+            },
+          },
+        },
+      });
+    });
+
+    it(`replaces the reference with null`, () => {
+      expect(snapshot.get(QueryRootId)).to.deep.eq({
+        foo: { id: 1, name: 'Foo', two: null },
+        bar: null,
+      });
+    });
+
+    it(`preserves nodes that only lost some of their inbound references`, () => {
+      expect(snapshot.get('1')).to.deep.eq({ id: 1, name: 'Foo', two: null });
+    });
+
+    it(`updates outbound references`, () => {
+      const queryRoot = snapshot.getNodeSnapshot(QueryRootId)!;
+      expect(queryRoot.outbound).to.have.deep.members([{ id: '1', path: ['foo'] }]);
+    });
+
+    it(`marks the container and all orphaned nodes as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members([QueryRootId, '1', '2', '111', '222', '333']);
+    });
+
+    it(`contains the correct nodes`, () => {
+      expect(snapshot.allNodeIds()).to.have.members([QueryRootId, '1']);
+    });
+
+  });
+});
