+++ test/unit/operations/write/writeValues.ts
@@ -1,0 +1,170 @@
+import * as util from 'util';
+
+import { CacheContext } from '../../../../src/context';
+import { GraphSnapshot } from '../../../../src/GraphSnapshot';
+import { EntitySnapshot } from '../../../../src/nodes';
+import { write } from '../../../../src/operations/write';
+import { NodeId, Query, StaticNodeId } from '../../../../src/schema';
+import { query } from '../../../helpers';
+
+const { QueryRoot: QueryRootId } = StaticNodeId;
+
+// These are really more like integration tests, given the underlying machinery.
+//
+// It just isn't very fruitful to unit test the individual steps of the write
+// workflow in isolation, given the contextual state that must be passed around.
+describe(`operations.write`, () => {
+  const config = new CacheContext({
+    logger: {
+      warn(message: string, ...args: any[]) {
+        throw new Error(util.format(message, ...args));
+      },
+    },
+  });
+  const rootValuesQuery = query(`{ foo bar }`);
+
+  const viewerQuery = query(`{
+    viewer {
+      id
+      name
+    }
+  }`);
+
+  const empty = new GraphSnapshot();
+
+  describe(`write only values to a root`, () => {
+    let snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      const result = write(config, empty, rootValuesQuery, { foo: 123, bar: 'asdf' });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`creates the query root, with the values`, () => {
+      expect(snapshot.get(QueryRootId)).to.deep.eq({ foo: 123, bar: 'asdf' });
+    });
+
+    it(`marks the root as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members([QueryRootId]);
+    });
+
+    it(`only contains the root node`, () => {
+      expect(snapshot.allNodeIds()).to.have.members([QueryRootId]);
+    });
+
+    it(`emits the root as an EntitySnapshot`, () => {
+      expect(snapshot.getNodeSnapshot(QueryRootId)).to.be.an.instanceOf(EntitySnapshot);
+    });
+  });
+
+  describe(`write a new single entity hanging off of a root`, () => {
+    let snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      const result = write(config, empty, viewerQuery, {
+        viewer: { id: 123, name: 'Gouda' },
+      });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`creates the query root, referencing the entity`, () => {
+      expect(snapshot.get(QueryRootId)).to.deep.eq({
+        viewer: { id: 123, name: 'Gouda' },
+      });
+    });
+
+    it(`indexes the entity`, () => {
+      expect(snapshot.get('123')).to.deep.eq({
+        id: 123, name: 'Gouda',
+      });
+    });
+
+    it(`emits the root as an EntitySnapshot`, () => {
+      expect(snapshot.getNodeSnapshot(QueryRootId)).to.be.an.instanceOf(EntitySnapshot);
+    });
+
+    it(`emits the entity as an EntitySnapshot`, () => {
+      expect(snapshot.getNodeSnapshot('123')).to.be.an.instanceOf(EntitySnapshot);
+    });
+
+    it(`directly references viewer from the query root`, () => {
+      const queryRoot = snapshot.get(QueryRootId);
+      const viewer = snapshot.get('123');
+      expect(queryRoot.viewer).to.eq(viewer);
+    });
+
+    it(`records the outbound reference from the query root`, () => {
+      const queryRoot = snapshot.getNodeSnapshot(QueryRootId)!;
+      expect(queryRoot.outbound).to.deep.eq([{ id: '123', path: ['viewer'] }]);
+      expect(queryRoot.inbound).to.eq(undefined);
+    });
+
+    it(`records the inbound reference from referenced entity`, () => {
+      const queryRoot = snapshot.getNodeSnapshot('123')!;
+      expect(queryRoot.inbound).to.deep.eq([{ id: QueryRootId, path: ['viewer'] }]);
+      expect(queryRoot.outbound).to.eq(undefined);
+    });
+
+    it(`marks the entity and root as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members([QueryRootId, '123']);
+    });
+
+    it(`only contains the two nodes`, () => {
+      expect(snapshot.allNodeIds()).to.have.members([QueryRootId, '123']);
+    });
+  });
+
+  describe(`write falsy values`, () => {
+    let falsyValuesQuery: Query, snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      falsyValuesQuery = query(`{ null, false, zero, string }`);
+
+      const result = write(config, empty, rootValuesQuery, { null: null, false: false, zero: 0, string: '' });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`persists all falsy values`, () => {
+      expect(snapshot.get(QueryRootId)).to.deep.eq({ null: null, false: false, zero: 0, string: '' });
+    });
+  });
+
+  describe(`write to inner nodes`, () => {
+
+    let baseline: GraphSnapshot, snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      const baselineResult = write(config, empty, rootValuesQuery, {
+        foo: { id: 1, name: 'Foo' },
+        bar: { id: 2, name: 'Bar' },
+      });
+      baseline = baselineResult.snapshot;
+
+      const innerNodeQuery = query(`{ name extra }`, undefined, '1');
+      const result = write(config, baseline, innerNodeQuery, {
+        name: 'moo',
+        extra: true,
+      });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`doesn't mutate the previous versions`, () => {
+      expect(baseline.get(QueryRootId)).to.not.eq(snapshot.get(QueryRootId));
+      expect(baseline.get('1')).to.not.eq(snapshot.get('1'));
+      expect(baseline.get('2')).to.eq(snapshot.get('2'));
+      expect(baseline.get(QueryRootId)).to.deep.eq({
+        foo: { id: 1, name: 'Foo' },
+        bar: { id: 2, name: 'Bar' },
+      });
+    });
+
+    it(`edits the inner node`, () => {
+      expect(snapshot.get('1')).to.deep.eq({ id: 1, name: 'moo', extra: true });
+    });
+
+    it(`marks only the inner node as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members(['1']);
+    });
+
+  });
+});
